Added new viewchange-request
As the leader is not polling anymore, the followers will not now when he's not available.
So the view-change request has been moved to AddTxRequest.
If the leader doesn't reply in 1 second, he is deemed faulty.
The node will then contact other nodes to try to contact the lader.
If enough nodes fail, a view-change will be initiated.